Iteracao sobre edges.json funcionando

+++ core/grafo.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 Grafo::Grafo(int vertices) : adjLista(vertices) {}
 
-void Grafo::addAresta(int origem, int destino, double peso) {
+void Grafo::addAresta(long origem, long destino, double peso) {
     //Adiciona aresta no vértice passado como origem
     adjLista[origem].push_back({destino, peso});
 }
